Fix bug in EphemeralRandomConstantNode
Currently, the `GetOutput` function returns a single constant for the node. This typically goes well, since this result is multiplied with other arrays from other nodes, but when the constant is the only node in the function, you get only a one-dimensional output. It should have the shape (length) of `X`, instead. (We came across this by accident, where we encountered the rare case where this went wrong.) This PR fixes this.